Repo Init (Capture)
Not completely sure about the history of this package, but I'm making a repo for it so it can be fully tracked, as the original one appears to have been hidden or deleted.

https://npm.im/@geekcornernpm/jimp-plugins

This seems to be the only remnant left of the source for it, so I wanted to make sure to track it, rather than only having the npm of it. I also want to modernize it, by making it type-safe and add type definitions support for use with TypeScript. For reference, it's in relation to the implementation for this project:

https://github.com/ModifiedCommand/ConvertJavaTextureToBedrock

I just ran this to get the source from npm:

```sh
npm pack @geekcornernpm/jimp-plugins
```